python/requirements.txt
Flask
Flask_Cors
PyYAML
